Refactor validator structure
